% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSurvivalData.R
\name{bootstrap_narlal}
\alias{bootstrap_narlal}
\title{Bootstraping of all survival data of the NARLAL trial}
\usage{
bootstrap_narlal(df, formcox, formsurv, nboot = 10, conf.int = 0.95, seed = 42)
}
\arguments{
\item{df}{data frame}

\item{formcox}{list of formula for Cox call}

\item{formsurv}{list of formula for survival call}

\item{nboot}{number of bootstraps (nboot=1 will replace the bootstrap values with the standard values)}

\item{conf.int}{confidence intervals extracted from the bootstrap. Will typpical be 0.95}

\item{seed}{a value to seed the random generator}
}
\value{
return a list containing all the standard acall and the related bootstrap values
}
\description{
Bootstraping of all survival data of the NARLAL trial
}
\keyword{internal}
