% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportPlotFunctions.R
\name{custom_tab_Narlal}
\alias{custom_tab_Narlal}
\title{custom_tab_Narlal}
\usage{
custom_tab_Narlal(df, header, footer = c())
}
\arguments{
\item{df}{the data.frame containing the table}

\item{header}{the text that will appear as header of the table}

\item{footer}{the text that will appear as footer of the table}
}
\value{
a flextable containing the table
}
\description{
An internal function used to define the layout of the tables generated by the package
}
\keyword{internal}
