% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportPlotFunctions.R
\name{customtab_defaults_Narlal}
\alias{customtab_defaults_Narlal}
\title{customtab_defaults_Narlal}
\usage{
customtab_defaults_Narlal()
}
\value{
no direct returned values. Only setting flextable defaults
}
\description{
An internal function used to define the defaults for the flextable used to produce the table of the package
}
\keyword{internal}
