% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoData.R
\name{DemoData}
\alias{DemoData}
\title{Create demo data fro the Narlal2 clinical trial}
\usage{
DemoData(nrpts = 360, uselabels = FALSE, seed = 42, firstid = 1, filepath)
}
\arguments{
\item{nrpts}{pt number to include in the demo data}

\item{uselabels}{can be set to TRUE if the values in the data table should be the label names (e.g. Male/Female versus 1/2). FALSE is the default and is needed if the data should be imported into a REdcap database}

\item{seed}{is used for the random number generator. Two runs will for a fixed value of the seed produce the same demo data}

\item{firstid}{is the id of the first patient, and the following are consecutive. It can be used to divide a large cohort into subfiles which might be needed in upload to Redcap that seems to have a limit on the number of patients that can be imported (likely due to memory issues at the Redcap server)}

\item{filepath}{path to the output file}
}
\value{
write a csv file with demo data that can be loaded into a redcap database
}
\description{
‘Demodata’ writes a CSV file to disk containing demo data for the clinical trial Narlal2. The file can be imported to the Redcap data database used for the trial. The primary purpose of the demo data has been to develop the other functionalities utilising “realistic” data.
}
\examples{
file <- file.path(tempdir(),'demo.csv')
DemoData(nrpts=100,uselabels=FALSE,seed=17,filepath=file)
}
