% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractImagingRecurenceDates.R
\name{ExtractImagingRecurrenceDates}
\alias{ExtractImagingRecurrenceDates}
\title{Extract recurrence dates from the medical image information}
\usage{
ExtractImagingRecurrenceDates(df)
}
\arguments{
\item{df}{the data.frame obtained from LoadAndPrepareData}
}
\value{
data.frame that contain recurrence dates
}
\description{
‘ExtractImagingRecurrenceDates’ extracts the dates of recurrence from the medical images. The dates are local recurrence or any recurrence.
}
\examples{
 file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
RecurrenceDates <- ExtractImagingRecurrenceDates(df)
}
