% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractPatientCharacteristic.R
\name{ExtractPatientCharacteristic}
\alias{ExtractPatientCharacteristic}
\title{Extract patient characteristics}
\usage{
ExtractPatientCharacteristic(rawdata)
}
\arguments{
\item{rawdata}{the data.frame obtained from LoadAndPrepareData}
}
\value{
data.frame that contains patient characteristic
}
\description{
'ExtracPatientCharacteristic' extract the patient characteristics from the initial raw data obtained using LoadAndPrepareData
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtChar <- ExtractPatientCharacteristic(df)
}
