% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSurvivalData.R
\name{ExtractSurvivalData}
\alias{ExtractSurvivalData}
\title{Extract survival data}
\usage{
ExtractSurvivalData(rawdata, censor_after_rt)
}
\arguments{
\item{rawdata}{the data.frame obtained from LoadAndPrepareData}

\item{censor_after_rt}{value in month of the latest value to censor after RT. Needs to be fixed then the database has a censor date included}
}
\value{
data.frame that contain survival data
}
\description{
'ExtracSurvivalData' extract the survival data from the initial raw data obtained using LoadAndPrepareData
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtSurv <- ExtractSurvivalData(df,12*5)
}
