% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportPlotFunctions.R
\name{ggcompetingrisks.cuminc_narlal}
\alias{ggcompetingrisks.cuminc_narlal}
\title{Copy of ggcompetingrisks.cuminc but need to enable edition of a two layout (order of plotting and conf.int usinng lines)}
\usage{
ggcompetingrisks.cuminc_narlal(
  fit,
  gnames = NULL,
  gsep = " ",
  multiple_panels = TRUE,
  coef = 1.96,
  conf.int = FALSE
)
}
\arguments{
\item{fit}{see ggcompetingrisks {survminer}}

\item{gnames}{see ggcompetingrisks {survminer}}

\item{gsep}{see ggcompetingrisks {survminer}}

\item{multiple_panels}{see ggcompetingrisks {survminer}}

\item{coef}{see ggcompetingrisks {survminer}}

\item{conf.int}{see ggcompetingrisks {survminer}
@keywords internal}
}
\value{
Returns an object of class gg.
}
\description{
Copy of ggcompetingrisks.cuminc but need to enable edition of a two layout (order of plotting and conf.int usinng lines)
}
