% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportPlotFunctions.R
\name{ggcompetingrisks_narlal}
\alias{ggcompetingrisks_narlal}
\title{Copy of ggcompetingrisks but needed to enable correction in a subfunction}
\usage{
ggcompetingrisks_narlal(
  fit,
  gnames = NULL,
  gsep = " ",
  multiple_panels = TRUE,
  ggtheme = survminer::theme_survminer(),
  coef = 1.96,
  conf.int = FALSE,
  ...
)
}
\arguments{
\item{fit}{see ggcompetingrisks {survminer}}

\item{gnames}{see ggcompetingrisks {survminer}}

\item{gsep}{see ggcompetingrisks {survminer}}

\item{multiple_panels}{see ggcompetingrisks {survminer}}

\item{ggtheme}{see ggcompetingrisks {survminer}}

\item{coef}{see ggcompetingrisks {survminer}}

\item{conf.int}{see ggcompetingrisks {survminer}}

\item{...}{see ggcompetingrisks {survminer}}
}
\value{
Returns an object of class gg.
}
\description{
Copy of ggcompetingrisks but needed to enable correction in a subfunction
}
\keyword{internal}
