% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAndPrepareData.R
\name{LoadAndPrepareData}
\alias{LoadAndPrepareData}
\title{Load csv file from the clinical trial Narlal2}
\usage{
LoadAndPrepareData(filename = "ExternalData/DemoData.csv")
}
\arguments{
\item{filename}{path to the cvs file that contains all the data from the NARLAL datbase}
}
\value{
data.frame that contain all thhe data from the NARLAL database
}
\description{
'LoadAndPrepareData'reads the data exported from the Narlal database (e.g. the included demo data)
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
}
