% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEarlyToxTables.R
\name{PlotEarlyToxTables}
\alias{PlotEarlyToxTables}
\title{Plot early toxicity tables}
\usage{
PlotEarlyToxTables(
  rawdata,
  maxMonth = 6,
  nBootImputation = 20,
  p_values = NULL,
  p_value_save_path = NULL,
  Q = 0.1
)
}
\arguments{
\item{rawdata}{The input data frame containing toxicity data. The data is the
output from the function ExtractToxicityData}

\item{maxMonth}{The maximum month that will contribute data for the after RT
values (See details - default is 6)}

\item{nBootImputation}{The number of multiple imputations that are performed
when calculating the p-values (default is 20)}

\item{p_values}{calculated from a previous run (can be helpful to store
p_values since they take quite a long time to calculate; if provided, the
p_values will not be recalculated; default is an empty string which will
trigger p-value calculation)}

\item{p_value_save_path}{a path to save the calculated p_values. The default
is an empty string that will prevent the p_values from being saved.}

\item{Q}{is the false discovery rate used to compensate for multiple testing.
The multiple testing procedure is based on Benjamini Y, Hochberg Y.
Controlling the False Discovery Rate: A Practical and Powerful Approach to
Multiple Testing. Journal of the Royal Statistical Society Series B
(Methodological). 1995;57:289-300. (the default value is 0.1)}
}
\value{
A flextable that can be saved as a Word file
}
\description{
Plot early toxicity tables
}
\details{
This function produces the table values needed to compare the
toxicity of the two treatment arms for the Early toxicity article. Details
of how the presented p_values are calculated can be found in the function
CalculatePVaulesEarlyTox. The function returns a flextable, which then can
be modified or saved directly as a Word file.
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtTox <- ExtractToxicityData(df)
index<-grepl('_fu_mdr_',names(PtTox)) | grepl('_rt_uge_',names(PtTox))
index<-index & !(grepl('diaria_fu_mdr_',names(PtTox)) |
   grepl('diaria_rt_uge_',names(PtTox)))
PtTox[,index]<-NULL
ft<-PlotEarlyToxTables(PtTox,nBootImputation=1)
flextable::save_as_docx(ft, path = 'c:/home/cab/temp/EarlyToxTable.docx',
  pr_section =  officer::prop_section(page_size =
  officer::page_size(orient = "portrait"), type = "continuous"))
}
