% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGradeTwoAndThreeEarlyTox.R
\name{PlotGradeTwoAndThreeEarlyTox}
\alias{PlotGradeTwoAndThreeEarlyTox}
\title{Plot grade two and three early toxicity}
\usage{
PlotGradeTwoAndThreeEarlyTox(rawdata, outputDirectory = "c:/home/cab/temp")
}
\arguments{
\item{rawdata}{The input data frame containing toxicity data. The data is the
output from the function ExtractToxicityData}

\item{outputDirectory}{is a path to where the plots are saved}
}
\value{
The output is the saved plots, and no R-data is returned
}
\description{
Plot grade two and three early toxicity
}
\details{
The function produces the time-dependent toxicity graphs for the early toxicity paper of the Narlal2 trial.
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtTox <- ExtractToxicityData(df)
PlotGradeTwoAndThreeEarlyTox(PtTox,outputDirectory='c:/home/cab/temp')
}
