% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPatientCharacteristic.R
\name{PlotPatientCharacteristic}
\alias{PlotPatientCharacteristic}
\title{Write word files containing patient characteristics}
\usage{
PlotPatientCharacteristic(df, filepath)
}
\arguments{
\item{df}{the output from the the function ExtracPatientCharacteristic}

\item{filepath}{the filepath to the directory where the patient characteristics shall be stored}
}
\value{
Is returning word files with tables of patient characteristics data
}
\description{
Write word files containing patient characteristics
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtChar <- ExtractPatientCharacteristic(df)
PlotPatientCharacteristic(df=PtChar,filepath='c:/home/cab/temp')
}
