% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSurvivalData.R
\name{PlotSurvivalData}
\alias{PlotSurvivalData}
\title{Creates survival plots}
\usage{
PlotSurvivalData(
  df,
  filepath,
  nboot = 10,
  conf.int = 0.95,
  seed = 42,
  ChangeText = c()
)
}
\arguments{
\item{df}{the output from the the function ExtractSurvivalData}

\item{filepath}{the filepath to the directory where the patient survival data shall be stored}

\item{nboot}{number of bootstraps (nboot=1 will replace the bootstrap values with the standard values)}

\item{conf.int}{confidence intervals extracted from the bootstrap. Will typical be 0.95}

\item{seed}{a value to seed the random generator}

\item{ChangeText}{is a variable used to change the text on plots just before they are plotted. The variable is defined in three part (see example below) that is used to change the text in variable/names, levels, and labels on the plots.}
}
\value{
Is exporting survival figures
}
\description{
Creates survival plots divided by treatment arm of local control, progression-free survival, and overall survival. Furthermore, a competing risk plot of the first event among local failure, distant failure and death is also created – the competing risk plots are made in two versions “standard”, including confidence intervals and a stacked version. For all plots, three versions are created: one for all patients and two for patients with tumours of histology squamous or non-squamous.
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtSurvival <- ExtractSurvivalData(df)
ChangeText<-c()
ChangeText$ChangeLabels <- c(
  `Time [months]`='Time since randomisation [Months]'
)
ChangeText$ChangeVar <-c(
  arm='Treatment Arm'
)
ChangeText$ChangeLevels <- c(
  Standard='Standard',
  Eskaleret='Escalated'
)
PlotSurvivalData(df=PtSurvival,filepath='c:/home/cab/temp',
                 nboot=10,conf.int=.95,seed=42,ChangeText=ChangeText)
}
