% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotToxicityData.R
\name{PlotToxicityData}
\alias{PlotToxicityData}
\title{Write word files containing patient toxicity data}
\usage{
PlotToxicityData(df, filepath)
}
\arguments{
\item{df}{the output from the the function ExtractToxicityData}

\item{filepath}{the filepath to the directory where the patient toxicity shall be stored}
}
\value{
Is returning word files with tables of patient toxicity data
}
\description{
Write word files containing patient toxicity data
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtTox <- ExtractToxicityData(df)
PlotToxicityData(df=PtTox,filepath='c:/home/cab/temp')
}
