% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractSurvivalData.R
\name{ExtractSurvivalData}
\alias{ExtractSurvivalData}
\title{Extract survival data}
\usage{
ExtractSurvivalData(rawdata, maxFollowUpTime = Inf)
}
\arguments{
\item{rawdata}{the data.frame obtained from LoadAndPrepareData}

\item{maxFollowUpTime}{time in months, after which all data are censored. Default time is infinity}
}
\value{
data.frame that contain survival data
}
\description{
'ExtracSurvivalData' extract the survival data from the initial raw data obtained using LoadAndPrepareData
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtSurv <- ExtractSurvivalData(df)
}
