% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractToxicityData.R
\name{ExtractToxicityData}
\alias{ExtractToxicityData}
\title{Extract toxicity data}
\usage{
ExtractToxicityData(rawdata)
}
\arguments{
\item{rawdata}{the data.frame obtained from LoadAndPrepareData}
}
\value{
data.frame that contain toxicity data
}
\description{
'ExtracToxicityData' extract the patient toxicity from the initial raw data obtained using LoadAndPrepareData
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtTox <- ExtractToxicityData(df)
}
