% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSurvivalData.R
\name{PlotSurvivalData}
\alias{PlotSurvivalData}
\title{Write word files containing patient toxicity data}
\usage{
PlotSurvivalData(df, filepath, nboot = 10, conf.int = 0.95, seed = 42)
}
\arguments{
\item{df}{the output from the the function ExtractSurvivalData}

\item{filepath}{the filepath to the directory where the patient survival data shall be stored}

\item{nboot}{number of bootstraps (nboot=1 will replace the bootstrap values with the standard values)}

\item{conf.int}{confidence intervals extracted from the bootstrap. Will typpical be 0.95}

\item{seed}{a value to seed the random generator}
}
\value{
Is exporting survival figures
}
\description{
Write word files containing patient toxicity data
}
\examples{
file <- system.file('extdata','DemoData.csv',package="Narlal2")
df <- LoadAndPrepareData(filename=file)
PtSurvival <- ExtractSurvivalData(df,12*5)
PlotSurvivalData(df=PtSurvival,filepath='c:/home/cab/temp',nboot=10,conf.int=.95,seed=42)
}
